# apps/api/src/treasury-transaction-cost.controller.ts
import{Body,Controller,Get,Param,Post,UseGuards}from"@nestjs/common";import{AuthContext}from"./auth.service";import{IdempotentCommand}from"./idempotency.interceptor";import{activeOrganizationId,CurrentAuth,ORGANIZATION_ROLES,RequireRoles,SessionGuard}from"./session.guard";import{CreateTreasuryTransactionCostDto}from"./treasury-transaction-cost.dto";import{TreasuryTransactionCostService}from"./treasury-transaction-cost.service";
@Controller("treasury-outcomes/:outcomeId/transaction-costs")@UseGuards(SessionGuard)@RequireRoles(...ORGANIZATION_ROLES)export class TreasuryTransactionCostController{constructor(private readonly service:TreasuryTransactionCostService){}@Get()list(@CurrentAuth()auth:AuthContext,@Param("outcomeId")outcomeId:string){return this.service.list(activeOrganizationId(auth),outcomeId)}@Get(":costId")get(@CurrentAuth()auth:AuthContext,@Param("outcomeId")outcomeId:string,@Param("costId")costId:string){return this.service.get(activeOrganizationId(auth),outcomeId,costId)}@Post()@RequireRoles("ADMIN","TREASURY_COMMITTEE","REVIEWER")@IdempotentCommand()create(@CurrentAuth()auth:AuthContext,@Param("outcomeId")outcomeId:string,@Body()body:CreateTreasuryTransactionCostDto){return this.service.create(activeOrganizationId(auth),outcomeId,body,auth.userId)}}