# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AEOS API
  version: 0.1.0
  description: Every response returns a validated or server-generated X-Request-ID that also links structured logs and new audit events. Cookie-authenticated unsafe requests require an allowed Origin and the session-bound X-CSRF-Token returned by auth/verify. Domain command endpoints accept Idempotency-Key and preserve the first completed response for 24 hours; reuse with different input returns 409.
servers: [{ url: http://localhost:4000/api/v1 }]
security: [{ cookieAuth: [] }]
paths:
...
  /evidence:
    get:
      summary: Search organization-scoped verified evidence
      parameters:
        - { in: query, name: status, schema: { enum: [VERIFIED, REJECTED, UNVERIFIED] } }
        - { in: query, name: freshness, schema: { enum: [FRESH, STALE, ARCHIVED] } }
        - { in: query, name: chainId, schema: { type: integer } }
        - { in: query, name: predicate, schema: { type: string } }
        - { in: query, name: minQuality, schema: { type: integer, minimum: 0, maximum: 100 } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { in: query, name: cursor, schema: { type: string } }
      responses: { '200': { description: Cursor page of Evidence } }
  /evidence/{id}:
    get:
      summary: Get Evidence without leaking cross-organization existence
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses: { '200': { description: Evidence }, '404': { description: Not found } }
...
  /treasury-outcomes/{outcomeId}/transaction-costs:
    get:
      summary: List immutable Evidence-bound observed transaction-cost assessments for one execution-linked Treasury outcome
      responses: { '200': { description: Cost assessments with counterfactual, net benefit, causality, automatic learning and asset authority withheld }, '404': { description: Parent outcome not found without cross-tenant leakage } }
    post:
      summary: Bind verified historical cost Evidence to the exact confirmed Safe transaction without claiming performance or causality
      parameters: [{ in: header, name: Idempotency-Key, required: false, schema: { type: string, maxLength: 200 } }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [evidenceSnapshotId], properties: { evidenceSnapshotId: { type: string } } } } }
      responses: { '201': { description: Immutable common-numeraire network fee, protocol fee and execution-shortfall total; slippage and price impact remain non-additive diagnostics }, '400': { description: Incomplete execution identity, unverified/conflicting/low-quality Evidence, component overlap, non-observed input, methodology, execution or numeraire mismatch }, '404': { description: Execution-linked outcome or Evidence snapshot not found without cross-tenant leakage } }
  /treasury-outcomes/{outcomeId}/transaction-costs/{costId}:
    get:
      summary: Read one immutable transaction-cost assessment and its exact Evidence, Safe, Policy and outcome lineage
      responses: { '200': { description: Frozen observed cost assessment and stable hash }, '404': { description: Not found without cross-tenant leakage } }
...
  /proposals:
    get:
      summary: List immutable organization-scoped governance proposal drafts
      responses: { '200': { description: Proposal drafts } }
    post:
      summary: Build a deterministic ERC20 proposal draft from an approved Decision and passing Simulation
      responses: { '201': { description: Immutable non-executable proposal draft with decoded calldata }, '400': { description: Approval, simulation, snapshot, policy, action, or calldata consistency guard failed }, '404': { description: Referenced organization-scoped object not found } }
  /proposals/{id}:
    get:
      summary: Read a proposal content hash, provenance chain, decoded action, and transaction arrays
      responses: { '200': { description: Proposal draft }, '404': { description: Not found without cross-tenant leakage } }
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: aeos_session
...